+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - Add ov-auto-complete AngularJS directive to build a text field with auto-completion
 - Add new CSS classes on table filters to distinguish filters displayed in the first line from the following filters
 - Add "autoComplete" filter type to the tableContainer API. "filterBy" property can now accept a filter of type "autoComplete" to display a text field with suggestions
+- Add "autoComplete", "editableAutoComplete", "horizontalAutoComplete" and "horizontalEditableAutoComplete" formly types based on ov-auto-complete AngularJS directive
 
 # 6.0.0 / 2018-10-17